Split DependencyUnresolvedModuleIntegrationTest
